Revert "Select dashboard"
