+++ app/src/main/java/practicaiufragments/dam/com/netbeast/CustomListAdapter.java
@@ -18,8 +18,6 @@ import com.android.volley.VolleyError;
 import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.JsonObjectRequest;
 
-import org.json.JSONArray;
-import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.HashMap;
@@ -36,7 +34,6 @@ public class CustomListAdapter extends BaseAdapter {
     private String IP;
     private String port;
     private String url;
-    private List<String> items;
 
 
     private HashMap<String, String> mRequestParams;
@@ -49,26 +46,14 @@ public class CustomListAdapter extends BaseAdapter {
         this.title = title;
     }
 
-    public CustomListAdapter(Activity activity, List<String> items) {
-        this.activity = activity;
-        this.appItems = null;
-        this.items = items;
-    }
-
     @Override
     public int getCount() {
-        if (appItems!=null)
-            return appItems.size();
-        else
-            return items.size();
+        return appItems.size();
     }
 
     @Override
     public Object getItem(int location) {
-        if (appItems!=null)
-            return appItems.get(location);
-        else
-            return items.get(location);
+        return appItems.get(location);
     }
 
     @Override
@@ -77,151 +62,129 @@ public class CustomListAdapter extends BaseAdapter {
     }
 
     @Override
-    public View getView(final int position, View convertView, ViewGroup parent) {
-        if (appItems!=null) {
-            if (inflater == null)
-                inflater = (LayoutInflater) activity
-                        .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-            if (convertView == null)
-                convertView = inflater.inflate(R.layout.custom_row, null);
+    public View getView(int position, View convertView, ViewGroup parent) {
 
-            ImageButton imgBt = (ImageButton) convertView.findViewById(R.id.imbutton);
-            TextView txt = (TextView) convertView.findViewById(R.id.text);
-            ImageButton bt = (ImageButton) convertView.findViewById(R.id.button);
+        if (inflater == null)
+            inflater = (LayoutInflater) activity
+                    .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        if (convertView == null)
+            convertView = inflater.inflate(R.layout.custom_row, null);
 
-            // getting app data for the row
-            final App app = appItems.get(position);
+        ImageButton imgBt = (ImageButton) convertView.findViewById(R.id.imbutton);
+        TextView txt = (TextView) convertView.findViewById(R.id.text);
+        ImageButton bt = (ImageButton) convertView.findViewById(R.id.button);
 
-            if (app.getLogoURL() != null)
-                if (app.getLogoBitmap() == null)
-                    // If the logo has not been downloaded yet...
-                    // download app logo and set it to image button
-                    new DownloadImageTask(imgBt, app).execute(app.getLogoURL());
-                else
-                    // If the logo has already been downloaded...
-                    // set it to image button
-                    imgBt.setImageBitmap(app.getLogoBitmap());
+        // getting app data for the row
+        final App app = appItems.get(position);
+
+        if (app.getLogoURL() != null)
+            if (app.getLogoBitmap() == null)
+                // If the logo has not been downloaded yet...
+                // download app logo and set it to image button
+                new DownloadImageTask(imgBt, app).execute(app.getLogoURL());
             else
-                imgBt.setImageResource(R.drawable.dflt);
+                // If the logo has already been downloaded...
+                // set it to image button
+                imgBt.setImageBitmap(app.getLogoBitmap());
+        else
+            imgBt.setImageResource(R.drawable.dflt);
 
-            // set app name to textView
-            txt.setText(app.getName());
+        // set app name to textView
+        txt.setText(app.getName());
 
 
-            // Let's create/get global params
-            Global g = Global.getInstance();
-            IP = g.getIP();
-            port = g.getPort();
+        // Let's create/get global params
+        Global g = Global.getInstance();
+        IP = g.getIP();
+        port = g.getPort();
 
-            mRequestParams = new HashMap<>();
+        mRequestParams = new HashMap<>();
 
-            // Set the button over the apps, to INSTALL, LAUNCH, REMOVE or STOP
-            switch (title) {
-                case "Install":
-                    bt.setVisibility(View.VISIBLE);
-                    if (app.getInstalled()) {
-                        bt.setImageResource(R.drawable.launch);
-                        bt.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View v) {
-                                url = "http://" + IP + ":" + port + "/api/activities/" + app.getName();
-                                // Use this url to post params
-                                mRequestParams.put("app", app.getName());
-                                // Make post request
-                                sendPostRequest();
-                                launchWebActivity(v, app.getName());
-                            }
-                        });
-                    } else {
-                        bt.setImageResource(R.drawable.install);
-                        bt.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View v) {
-                                url = "http://" + IP + ":" + port + "/api/apps";
-                                String gitUrl = "https://github.com/" + app.getFull_name();
-                                // Use this url to post params
-                                mRequestParams.put("url", gitUrl);
-                                // Make post request
-                                sendPostRequest();
-                            }
-                        });
-                    }
-                    break;
-                case "Activities":
-                    bt.setVisibility(View.VISIBLE);
-                    bt.setImageResource(R.drawable.stop);
+        // Set the button over the apps, to INSTALL, LAUNCH, REMOVE or STOP
+        switch(title){
+            case "Install":
+                bt.setVisibility(View.VISIBLE);
+                if (app.getInstalled()) {
+                    bt.setImageResource(R.drawable.launch);
                     bt.setOnClickListener(new View.OnClickListener() {
                         @Override
                         public void onClick(View v) {
                             url = "http://" + IP + ":" + port + "/api/activities/" + app.getName();
+                            //String gitUrl = "https://github.com/" + app.getFull_name();
                             // Use this url to post params
-                            mRequestParams.put("url", url);
+                            mRequestParams.put("app", app.getName());
                             // Make post request
-                            sendDeleteRequest();
+                            sendPostRequest();
+                            launchWebActivity(v, app.getName());
                         }
                     });
-                    imgBt.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-
-                        }
-                    });
-                    break;
-                case "Remove":
-                    bt.setVisibility(View.VISIBLE);
-                    bt.setImageResource(R.drawable.remove);
+                }
+                else {
+                    bt.setImageResource(R.drawable.install);
                     bt.setOnClickListener(new View.OnClickListener() {
                         @Override
                         public void onClick(View v) {
-                            url = "http://" + IP + ":" + port + "/api/apps/" + app.getName();
+
+                            url = "http://" + IP + ":" + port + "/api/apps";
+                            String gitUrl = "https://github.com/" + app.getFull_name();
                             // Use this url to post params
-                            mRequestParams.put("url", url);
+                            mRequestParams.put("url", gitUrl);
                             // Make post request
-                            sendDeleteRequest();
+                            sendPostRequest();
+
                         }
                     });
-                    break;
-            }
-
-            // If you click on an app, it launches
-            imgBt.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    url = "http://" + IP + ":" + port + "/api/activities/" + app.getName();
-                    // Use the app name to post params
-                    mRequestParams.put("app", app.getName());
-                    // Make post request
-                    sendPostRequest();
-                    launchWebActivity(v, app.getName());
                 }
-            });
-        }
-        else{
-            if (inflater == null)
-                inflater = (LayoutInflater) activity
-                        .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-            if (convertView == null)
-                convertView = inflater.inflate(R.layout.dashboard_row, null);
-
-            TextView txt = (TextView) convertView.findViewById(R.id.dashboard_text);
-            txt.setText(items.get(position));
-            txt.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    Global g = Global.getInstance();
-                    JSONArray dashboards = g.getDashboards();
-                    try {
-                        JSONObject dash = (JSONObject) dashboards.get(position);
-                        g.setIP(dash.getString("ip"));
-                        g.setPort(dash.getString("port"));
-                    } catch (JSONException e) {
-                        e.printStackTrace();
+                break;
+            case "Activities":
+                bt.setVisibility(View.VISIBLE);
+                bt.setImageResource(R.drawable.stop);
+                bt.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        url = "http://" + IP + ":" + port + "/api/activities/" + app.getName();
+                        // Use this url to post params
+                        mRequestParams.put("url", url);
+                        // Make post request
+                        sendDeleteRequest();
                     }
-                    Intent intent = new Intent(v.getContext(), MainActivity.class);
-                    v.getContext().startActivity(intent);
-                }
-            });
+                });
+                imgBt.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+
+                    }
+                });
+                break;
+            case "Remove":
+                bt.setVisibility(View.VISIBLE);
+                bt.setImageResource(R.drawable.remove);
+                bt.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        url = "http://" + IP + ":" + port + "/api/apps/" + app.getName();
+                        // Use this url to post params
+                        mRequestParams.put("url", url);
+                        // Make post request
+                        sendDeleteRequest();
+                    }
+                });
+                break;
         }
+
+        // If you click on an app, it launches
+        imgBt.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                url = "http://" + IP + ":" + port + "/api/activities/" + app.getName();
+                // Use the app name to post params
+                mRequestParams.put("app", app.getName());
+                // Make post request
+                sendPostRequest();
+                launchWebActivity(v, app.getName());
+            }
+        });
+
         return convertView;
     }
 
@@ -267,7 +230,6 @@ public class CustomListAdapter extends BaseAdapter {
         // Adding request to request queue
         QueueController.getInstance().addToRequestQueue(req);
     }
-
     public void launchWebActivity (View view, String name)
     {
         Intent intent = new Intent(view.getContext(), WebActivity.class);
