+++ app/src/main/java/practicaiufragments/dam/com/netbeast/GitInstallActivity.java
@@ -1,10 +1,9 @@
 package practicaiufragments.dam.com.netbeast;
 
+import android.app.Activity;
 import android.app.ProgressDialog;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
-import android.view.Menu;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -25,7 +24,7 @@ import java.util.TimerTask;
 /**
  * Created by Cayetano Rodríguez Medina on 29/4/16.
  */
-public class GitInstallActivity extends AppCompatActivity{
+public class GitInstallActivity extends Activity{
     private String IP;
     private String port;
     private String urlPostApp;
@@ -40,15 +39,10 @@ public class GitInstallActivity extends AppCompatActivity{
 
     private HashMap<String, String> mRequestParams;
 
-    private NavigationViewListener navigationViewListener;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.git_install_activity);
-
-        navigationViewListener = new NavigationViewListener(this);
-        getSupportActionBar().setTitle(R.string.install_activity_title);
 
         // Let's create/get global params
         Global g = Global.getInstance();
@@ -77,13 +71,6 @@ public class GitInstallActivity extends AppCompatActivity{
             }
         });
 
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
     }
 
     public void installApp() {
