+++ app/src/main/java/practicaiufragments/dam/com/netbeast/Global.java
@@ -1,9 +1,5 @@
 package practicaiufragments.dam.com.netbeast;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
 /**
  * Created by Cayetano Rodríguez Medina on 26/4/16.
  */
@@ -12,10 +8,9 @@ import org.json.JSONObject;
 public class Global {
     private static Global instance;
 
-    // IP and port where the dashboard is running
-    private String IP = "dashboard.827722d5.svc.dockerapp.io";
-    private String port = "";
-    private JSONArray dashboards;
+    // IP where the dashboard is running
+    private String IP = "192.168.1.138";
+    private String port = "8000";
 
     public void setIP (String ip) { this.IP = ip; }
 
@@ -24,28 +19,6 @@ public class Global {
     public void setPort (String port) { this.port = port; }
 
     public String getPort () { return this.port; }
-
-    public void addDashboard (String ip, String port) {
-        if (dashboards == null)
-            dashboards = new JSONArray();
-
-        JSONObject dash = new JSONObject();
-        try {
-            dash.put("ip", ip);
-            dash.put("port", port);
-        }catch (JSONException e){
-            e.printStackTrace();
-        }
-        dashboards.put(dash);
-    }
-
-    public JSONArray getDashboards() {
-        return this.dashboards;
-    }
-
-    public void clearDashboards() {
-        dashboards = null;
-    }
 
     // Singleton implementation
     public static synchronized Global getInstance() {
