+++ app/src/main/java/practicaiufragments/dam/com/netbeast/WebActivity.java
@@ -1,12 +1,12 @@
 package practicaiufragments.dam.com.netbeast;
 
+import android.app.Activity;
 import android.app.ProgressDialog;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
-import android.view.Menu;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
+import android.widget.TextView;
 
 import java.util.Timer;
 import java.util.TimerTask;
@@ -15,34 +15,33 @@ import java.util.TimerTask;
  * Created by Cayetano Rodríguez Medina on 12/5/16.
  */
 
-public class WebActivity extends AppCompatActivity {
-    private String app_name;
+public class WebActivity extends Activity {
+    private String title;
     private String IP;
     private String port;
 
     private static String TAG = WebActivity.class.getSimpleName();
 
-    private NavigationViewListener navigationViewListener;
-
     // Progress dialog
     private ProgressDialog pDialog;
 
+    private TextView textView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
         this.setContentView(R.layout.web_app);
 
-        // Get the app name to include it as the title of the activity
         Bundle b = getIntent().getExtras();
         if (b!=null) {
-            app_name = b.getString("title");
+            title = b.getString("title");
         }
         else
             Log.d(TAG, "Bundle is null");
 
-        navigationViewListener = new NavigationViewListener(this);
-        getSupportActionBar().setTitle(app_name);
+        textView = (TextView) findViewById(R.id.title_text);
+        textView.setText(title);
 
         // Let's create/get global params
         Global g = Global.getInstance();
@@ -62,7 +61,7 @@ public class WebActivity extends AppCompatActivity {
         showpDialog();
 
         // Load proxy app url
-        myWebView.loadUrl("http://" + IP + ":" + port + "/i/" + app_name);
+        myWebView.loadUrl("http://" + IP + ":" + port + "/i/" + title);
 
         new Timer().schedule(
                 new TimerTask() {
@@ -71,13 +70,8 @@ public class WebActivity extends AppCompatActivity {
                         hidepDialog();
                     }
                 }, 2000);
-    }
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
+
     }
 
     private void showpDialog() {
