--- app/src/main/res/layout-port/main_activity.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:background="@color/background"
-    android:id="@+id/t">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/connectedto"
-        android:id="@+id/tv_connectedto"
-        android:textColor="@color/text"
-        android:typeface="sans"
-        android:textSize="@dimen/connected"
-        android:layout_alignBottom="@+id/tv_dashboardip"
-        android:layout_toLeftOf="@+id/imageButton"
-        android:layout_toStartOf="@+id/imageButton" />
-
-    <ImageButton
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/imageButton"
-        android:src="@drawable/btn_explore_white"
-        android:background="@color/imgBackground"
-        android:onClick="exploreApps"
-        android:layout_above="@+id/tv_explore"
-        android:layout_centerHorizontal="true" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/explore"
-        android:id="@+id/tv_explore"
-        android:textColor="#ffffff"
-        android:typeface="sans"
-        android:textSize="@dimen/app_name"
-        android:layout_above="@+id/imageButton2"
-        android:layout_alignRight="@+id/tv_install"
-        android:layout_alignEnd="@+id/tv_install"
-        android:layout_marginBottom="47dp" />
-
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/none"
-        android:id="@+id/tv_dashboardip"
-        android:textColor="@color/error"
-        android:onClick="changeIp"
-        android:clickable="true"
-        android:layout_above="@+id/imageButton"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="37dp" />
-
-    <ImageButton
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/imageButton2"
-        android:src="@drawable/rsz_netbot_cajita_color_black_resized"
-        android:background="@color/imgBackground"
-        android:onClick="installApps"
-        android:layout_above="@+id/tv_install"
-        android:layout_toRightOf="@+id/tv_connectedto"
-        android:layout_toEndOf="@+id/tv_connectedto" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/install"
-        android:id="@+id/tv_install"
-        android:textColor="@color/text"
-        android:typeface="sans"
-        android:textSize="@dimen/app_name"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="49dp" />
-
-</RelativeLayout>
