+++ app/src/main/res/layout/custom_row.xml
@@ -8,7 +8,7 @@
     android:id="@+id/linearchild">
 
     <ImageButton
-        android:layout_width="175dp"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/imbutton"
         android:src="@drawable/dflt"
@@ -17,8 +17,7 @@
         android:layout_marginBottom="8dp"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
-        android:adjustViewBounds="true"
-        android:scaleType="fitXY" />
+        android:onClick=""/>
 
     <TextView
         android:layout_width="fill_parent"
@@ -35,15 +34,13 @@
         android:layout_marginBottom="10dp"/>
 
     <ImageButton
-        android:layout_width="125dp"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/button"
         android:visibility="invisible"
         android:src="@drawable/install"
         android:background="@color/imgBackground"
         android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true"
-        android:adjustViewBounds="true"
-        android:scaleType="fitXY"/>
+        android:layout_centerHorizontal="true" />
 
 </RelativeLayout>
