--- app/src/main/res/layout/dashboard_row.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content"
-    android:descendantFocusability="blocksDescendants"
-    android:orientation="vertical"
-    android:background="@color/background"
-    android:id="@+id/dashboardlinearchild">
-    <TextView
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/dashboard_text"
-        android:textColor="@color/colorAccent"
-        android:typeface="sans"
-        android:textSize="@dimen/app_name"
-        android:textAlignment="center"
-        android:text="Dashboard IP"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:layout_marginBottom="10dp"/>
-</RelativeLayout>
