--- app/src/main/res/layout/discovering_dashboards_splash.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:background="@color/background">
-
-
-    <TextView
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/discovering_text"
-        android:text="@string/discovering_dashboards"
-        android:textColor="@color/text"
-        android:textAlignment="center"
-        android:textSize="30sp"
-        android:typeface="sans"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"/>
-
-    <ProgressBar
-        style="@android:style/Widget.ProgressBar.Large"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/progressBar"
-        android:layout_below="@id/discovering_text"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="40dp"
-        android:indeterminate="true"/>
-
-</RelativeLayout>
