+++ app/src/main/res/layout/git_install_activity.xml
@@ -1,72 +1,72 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:app="http://schemas.android.com/apk/res-auto"
-xmlns:tools="http://schemas.android.com/tools"
-android:id="@+id/drawer_layout"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-android:fitsSystemWindows="true"
-tools:openDrawer="start">
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:fitsSystemWindows="true">
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="vertical"
+    android:background="@color/background">
 
-    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".ExploreActivity"
+    <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:background="@color/background">
+        android:layout_height="wrap_content"
+        android:theme="@style/AppTheme.AppBarOverlay"
+        android:id="@+id/view">
 
-        <android.support.design.widget.AppBarLayout
-            android:id="@+id/appBarLayout"
+        <android.support.v7.widget.Toolbar
+            android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:theme="@style/AppTheme.AppBarOverlay">
+            android:layout_height="?attr/actionBarSize"
+            android:background="?attr/colorPrimary"
+            app:popupTheme="@style/AppTheme.PopupOverlay">
 
-            <android.support.v7.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:background="?attr/colorPrimary"
-                app:popupTheme="@style/AppTheme.PopupOverlay" />
-
-        </android.support.design.widget.AppBarLayout>
-
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:layout_centerHorizontal="true"
-            android:layout_centerVertical="true"
-            android:id="@+id/linearLayout">
-
-            <EditText
-                android:id="@+id/editText"
+            <ImageView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:singleLine="true"
-                android:inputType="text"
+                android:src="@mipmap/logo"
+                android:paddingRight="10dp"
+                android:paddingBottom="8dp"/>
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:id="@+id/text"
+                android:textStyle="bold"
                 android:textColor="@color/text"
                 android:typeface="sans"
-                android:textSize="@dimen/app_name"
-                android:textAlignment="center"
-                android:layout_alignBaseline="@+id/installButton"
-                android:layout_alignBottom="@+id/installButton"
-                android:layout_centerHorizontal="true"
-                android:layout_weight="1"
-                android:hint="@string/hint_install_activity"
-                android:textColorHint="@color/hintText">
+                android:textSize="20sp"
+                android:text="@string/install_git"/>
+        </android.support.v7.widget.Toolbar>
 
-                <requestFocus />
+    </android.support.design.widget.AppBarLayout>
 
-            </EditText>
-        </LinearLayout>
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:layout_centerHorizontal="true"
+        android:layout_centerVertical="true"
+        android:id="@+id/linearLayout">
+
+        <EditText
+            android:id="@+id/editText"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:singleLine="true"
+            android:inputType="text"
+            android:textColor="@color/text"
+            android:typeface="sans"
+            android:textSize="@dimen/app_name"
+            android:textAlignment="center"
+            android:layout_alignBaseline="@+id/installButton"
+            android:layout_alignBottom="@+id/installButton"
+            android:layout_centerHorizontal="true"
+            android:layout_weight="1"
+            android:hint="@string/hint_install_activity"
+            android:textColorHint="@color/hintText">
+
+        <requestFocus />
+
+    </EditText>
+    </LinearLayout>
         <Button
             android:id="@+id/installButton"
             android:layout_width="wrap_content"
@@ -76,15 +76,6 @@ tools:openDrawer="start">
             android:textSize="@dimen/app_name"
             android:layout_below="@+id/linearLayout"
             android:layout_centerHorizontal="true" />
-    </RelativeLayout>
 
-</android.support.design.widget.CoordinatorLayout>
-<android.support.design.widget.NavigationView
-    android:id="@+id/nav_view"
-    android:layout_width="wrap_content"
-    android:layout_height="match_parent"
-    android:layout_gravity="start"
-    android:fitsSystemWindows="true"
-    app:headerLayout="@layout/nav_header_main"
-    app:menu="@menu/activity_main_drawer" />
-</android.support.v4.widget.DrawerLayout>
+
+</RelativeLayout>
