+++ app/src/main/res/layout/install_activity.xml
@@ -1,78 +1,68 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:app="http://schemas.android.com/apk/res-auto"
-xmlns:tools="http://schemas.android.com/tools"
-android:id="@+id/drawer_layout"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-android:fitsSystemWindows="true"
-tools:openDrawer="start">
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:fitsSystemWindows="true">
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="vertical"
+    android:background="@color/background">
 
-    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".ExploreActivity"
+    <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:background="@color/background">
+        android:layout_height="wrap_content"
+        android:theme="@style/AppTheme.AppBarOverlay"
+        android:id="@+id/view">
 
-        <android.support.design.widget.AppBarLayout
-            android:id="@+id/appBarLayout"
+        <android.support.v7.widget.Toolbar
+            android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:theme="@style/AppTheme.AppBarOverlay">
+            android:layout_height="?attr/actionBarSize"
+            android:background="?attr/colorPrimary"
+            app:popupTheme="@style/AppTheme.PopupOverlay">
 
-            <android.support.v7.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:background="?attr/colorPrimary"
-                app:popupTheme="@style/AppTheme.PopupOverlay" />
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@mipmap/logo"
+                android:paddingRight="10dp"
+                android:paddingBottom="8dp"/>
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:id="@+id/text"
+                android:textStyle="bold"
+                android:textColor="@color/text"
+                android:typeface="sans"
+                android:textSize="20sp"
+                android:text="@string/install_activity_title"/>
+        </android.support.v7.widget.Toolbar>
 
-        </android.support.design.widget.AppBarLayout>
+    </android.support.design.widget.AppBarLayout>
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/installButton1"
+        android:text="@string/install_explore"
+        android:textColor="@color/text"
+        android:typeface="sans"
+        android:textSize="@dimen/app_name"
+        android:textAlignment="center"
+        android:onClick="exploreInstallableApps"
+        android:layout_centerVertical="true"
+        android:layout_centerHorizontal="true" />
 
 
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/installButton1"
-            android:text="@string/install_explore"
-            android:textColor="@color/text"
-            android:typeface="sans"
-            android:textSize="@dimen/app_name"
-            android:textAlignment="center"
-            android:onClick="exploreInstallableApps"
-            android:layout_centerVertical="true"
-            android:layout_centerHorizontal="true" />
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/installButton2"
+        android:text="@string/install_git"
+        android:textColor="@color/text"
+        android:typeface="sans"
+        android:textAlignment="center"
+        android:textSize="@dimen/app_name"
+        android:onClick="gitInstallApps"
+        android:layout_below="@+id/installButton1"
+        android:layout_centerHorizontal="true" />
 
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/installButton2"
-            android:text="@string/install_git"
-            android:textColor="@color/text"
-            android:typeface="sans"
-            android:textAlignment="center"
-            android:textSize="@dimen/app_name"
-            android:onClick="gitInstallApps"
-            android:layout_below="@+id/installButton1"
-            android:layout_centerHorizontal="true" />
-    </RelativeLayout>
-
-</android.support.design.widget.CoordinatorLayout>
-<android.support.design.widget.NavigationView
-    android:id="@+id/nav_view"
-    android:layout_width="wrap_content"
-    android:layout_height="match_parent"
-    android:layout_gravity="start"
-    android:fitsSystemWindows="true"
-    app:headerLayout="@layout/nav_header_main"
-    app:menu="@menu/activity_main_drawer" />
-</android.support.v4.widget.DrawerLayout>
+</RelativeLayout>
