+++ app/src/main/res/layout/ip_dialog.xml
@@ -49,17 +49,5 @@
         android:hint="@string/hint_dialog_text"
         android:textColorHint="@color/hintText"
         android:textColor="@color/text"/>
-    <EditText
-        android:id="@+id/portTextEdit"
-        android:inputType="text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:layout_marginLeft="4dp"
-        android:layout_marginRight="4dp"
-        android:layout_marginBottom="4dp"
-        android:hint="@string/hint_port_dialog_text"
-        android:textColorHint="@color/hintText"
-        android:textColor="@color/text"/>
 
 </LinearLayout>
