+++ app/src/main/res/layout/main_activity.xml
@@ -2,7 +2,7 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="horizontal"
+    android:orientation="vertical"
     android:background="@color/background"
     android:id="@+id/t">
 
@@ -14,12 +14,9 @@
         android:textColor="@color/text"
         android:typeface="sans"
         android:textSize="@dimen/connected"
-        android:layout_marginLeft="44dp"
-        android:layout_marginTop="70dp"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:layout_marginStart="44dp" />
+        android:layout_alignBottom="@+id/tv_dashboardip"
+        android:layout_toLeftOf="@+id/imageButton"
+        android:layout_toStartOf="@+id/imageButton" />
 
     <ImageButton
         android:layout_width="wrap_content"
@@ -28,12 +25,8 @@
         android:src="@drawable/btn_explore_white"
         android:background="@color/imgBackground"
         android:onClick="exploreApps"
-        android:layout_marginLeft="51dp"
-        android:layout_marginStart="51dp"
-        android:layout_marginBottom="36dp"
-        android:layout_alignBottom="@+id/imageButton2"
-        android:layout_alignLeft="@+id/tv_connectedto"
-        android:layout_alignStart="@+id/tv_connectedto" />
+        android:layout_above="@+id/tv_explore"
+        android:layout_centerHorizontal="true" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -43,9 +36,10 @@
         android:textColor="#ffffff"
         android:typeface="sans"
         android:textSize="@dimen/app_name"
-        android:layout_alignTop="@+id/tv_install"
-        android:layout_toRightOf="@+id/tv_connectedto"
-        android:layout_toEndOf="@+id/tv_connectedto" />
+        android:layout_above="@+id/imageButton2"
+        android:layout_alignRight="@+id/tv_install"
+        android:layout_alignEnd="@+id/tv_install"
+        android:layout_marginBottom="47dp" />
 
 
     <TextView
@@ -56,11 +50,9 @@
         android:textColor="@color/error"
         android:onClick="changeIp"
         android:clickable="true"
-        android:layout_alignBottom="@+id/tv_connectedto"
-        android:layout_toRightOf="@+id/tv_connectedto"
-        android:layout_toEndOf="@+id/tv_connectedto"
-        android:layout_marginLeft="27dp"
-        android:layout_marginStart="27dp" />
+        android:layout_above="@+id/imageButton"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="37dp" />
 
     <ImageButton
         android:layout_width="wrap_content"
@@ -69,12 +61,9 @@
         android:src="@drawable/rsz_netbot_cajita_color_black_resized"
         android:background="@color/imgBackground"
         android:onClick="installApps"
-        android:layout_marginRight="66dp"
-        android:layout_marginEnd="66dp"
-        android:layout_marginBottom="10dp"
         android:layout_above="@+id/tv_install"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentEnd="true" />
+        android:layout_toRightOf="@+id/tv_connectedto"
+        android:layout_toEndOf="@+id/tv_connectedto" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -84,14 +73,9 @@
         android:textColor="@color/text"
         android:typeface="sans"
         android:textSize="@dimen/app_name"
-        android:layout_marginRight="44dp"
-        android:layout_marginEnd="44dp"
-        android:layout_marginBottom="25dp"
         android:layout_alignParentBottom="true"
-        android:layout_alignRight="@+id/imageButton2"
-        android:layout_alignEnd="@+id/imageButton2" />
-
-
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="49dp" />
 
 
 </RelativeLayout>
