--- app/src/main/res/layout/select_dashboard_activity.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:orientation="vertical"
-    android:background="@color/background">
-
-    <android.support.design.widget.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:theme="@style/AppTheme.AppBarOverlay"
-        android:id="@+id/view">
-
-        <android.support.v7.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
-            app:popupTheme="@style/AppTheme.PopupOverlay">
-
-            <ImageView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@mipmap/logo"
-                android:paddingRight="10dp"
-                android:paddingBottom="8dp"/>
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/text"
-                android:textStyle="bold"
-                android:textColor="@color/text"
-                android:typeface="sans"
-                android:textSize="20sp"
-                android:text="@string/select_dashboard"/>
-        </android.support.v7.widget.Toolbar>
-
-    </android.support.design.widget.AppBarLayout>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_dashboards"
-        android:text="@string/choose_dashboard"
-        android:textColor="@color/text"
-        android:typeface="sans"
-        android:textSize="25sp"
-        android:layout_below="@+id/view"
-        android:layout_centerHorizontal="true" />
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/installButton1"
-        android:text="Use dashboard in the cloud"
-        android:textColor="@color/text"
-        android:typeface="sans"
-        android:textSize="@dimen/app_name"
-        android:textAlignment="center"
-        android:onClick="cloudDashboard"
-        android:layout_below="@+id/list"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="50dp" />
-    <ListView
-        android:id="@+id/list"
-        android:layout_width="wrap_content"
-        android:layout_height="350dp"
-        android:layout_below="@id/tv_dashboards" />
-
-
-</RelativeLayout>
