--- app/src/main/res/layout/splash_layout.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:background="@color/colorPrimary">
-
-    <ImageView android:id="@+id/splashscreen"
-        android:layout_width="250dp"
-        android:layout_height="wrap_content"
-        android:src="@drawable/nb_logo"
-        android:layout_gravity="center"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:adjustViewBounds="true"
-        android:scaleType="fitXY"/>
-
-    <TextView android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/app_name"
-        android:textColor="@color/text"
-        android:textAlignment="center"
-        android:textSize="50sp"
-        android:typeface="sans"
-        android:layout_below="@id/splashscreen"/>
-
-</RelativeLayout>
